BLD: Numpy 1.20 compatibility

+++ src/porepy/fracs/fractures_3d.py
@@ -929,8 +929,8 @@ class FractureNetwork3d(object):
             # fractures. The indexing is a bit involved, but is based on there being
             # two intersection points for each segment - thus the indices in i0 and i1
             # must be divided by two.
-            on_bound_0 = bound_info[ind_0][np.floor(i0[0] / 2).astype(np.int)]
-            on_bound_1 = bound_info[ind_1][np.floor(i1[0] / 2).astype(np.int)]
+            on_bound_0 = bound_info[ind_0][np.floor(i0[0] / 2).astype(int)]
+            on_bound_1 = bound_info[ind_1][np.floor(i1[0] / 2).astype(int)]
 
             # Add the intersection to the internal list
             self.intersections.append(
@@ -1198,8 +1198,8 @@ class FractureNetwork3d(object):
         )
 
         # Update the edges_2_frac map to refer to the new edges
-        edges_2_frac_new = e_unique.shape[1] * [np.empty(0, dtype=np.int)]
-        is_boundary_edge_new = e_unique.shape[1] * [np.empty(0, dtype=np.int)]
+        edges_2_frac_new = e_unique.shape[1] * [np.empty(0, dtype=int)]
+        is_boundary_edge_new = e_unique.shape[1] * [np.empty(0, dtype=int)]
 
         for old_i, new_i in enumerate(all_2_unique_e):
             edges_2_frac_new[new_i], ind = np.unique(
@@ -1409,7 +1409,7 @@ class FractureNetwork3d(object):
         vertex or as internal.
 
         Returns:
-            list of np.int: indices of fractures, one list item per point.
+            list of int: indices of fractures, one list item per point.
 
         """
         fracs_of_points = []
@@ -1659,7 +1659,7 @@ class FractureNetwork3d(object):
         if inds.size > 0:
             split_frac = np.where(np.bincount(inds) > 1)[0]
         else:
-            split_frac = np.zeros(0, dtype=np.int)
+            split_frac = np.zeros(0, dtype=int)
 
         ind_map = np.delete(old_frac_ind, delete_frac)
 
@@ -1824,7 +1824,7 @@ class FractureNetwork3d(object):
 
         # ... on the points...
         point_tags = constants.NEUTRAL_TAG * np.ones(
-            self.decomposition["points"].shape[1], dtype=np.int
+            self.decomposition["points"].shape[1], dtype=int
         )
         # and the mapping between fractures and edges.
         edges_2_frac = self.decomposition["edges_2_frac"]
@@ -2305,7 +2305,7 @@ class FractureNetwork3d(object):
 
         """
         if constraints is None:
-            constraints = np.array([], dtype=np.int)
+            constraints = np.array([], dtype=int)
 
         # Extract geometrical information.
         p = self.decomposition["points"]
@@ -2337,8 +2337,8 @@ class FractureNetwork3d(object):
 
         # Count the number of times a lines is defined as 'inside' a fracture, and the
         # the number of times this is casued by a constraint
-        in_frac_occurences = np.zeros(edges.shape[1], dtype=np.int)
-        in_frac_occurences_by_constraints = np.zeros(edges.shape[1], dtype=np.int)
+        in_frac_occurences = np.zeros(edges.shape[1], dtype=int)
+        in_frac_occurences_by_constraints = np.zeros(edges.shape[1], dtype=int)
 
         for poly_ind, poly_edges in enumerate(self.decomposition["line_in_frac"]):
             for edge_ind in poly_edges:
@@ -2387,7 +2387,7 @@ class FractureNetwork3d(object):
         intersection_points = []
         for pi in intersection_point_canditates:
             _, edge_ind = np.where(edges == pi)
-            frac_arr = np.array([], dtype=np.int)
+            frac_arr = np.array([], dtype=int)
             for e in edge_ind:
                 frac_arr = np.append(frac_arr, self.decomposition["edges_2_frac"][e])
             unique_fracs = np.unique(frac_arr)
@@ -2599,7 +2599,7 @@ class FractureNetwork3d(object):
         # Intersection points, in 2d coordinates
         ip = np.empty((2, 0))
 
-        other_frac = np.empty(0, dtype=np.int)
+        other_frac = np.empty(0, dtype=int)
 
         for i in isect:
             if i.first.index == frac_num:
