+++ src/porepy/utils/sparse_mat.py
@@ -411,7 +411,7 @@ def _csx_matrix_from_blocks(
         )
         indices = base + block_increase
     else:
-        indices = np.arange(num_blocks, dytpe=np.int)
+        indices = np.arange(num_blocks, dytpe=int)
 
     mat = matrix_format(
         (data, indices, indptr),
